Make 144x144 data matrix readable
According to this bug (opened in the java version of the library, but ported to the CPP version)
the 144x144 datamatrices are somehow special in the standard, and require special handling while decoding.
However, it seems that this special handling was implemented in the decodign code, but not in the encoding code.
Therefore, while ZXing seems to be able to decode standard 144x144 datamatrices generated by other tools,
it cannot decode its own 144x144 datamatrices.

Since I cannot find how tho change the encoding part to agree with the standard,
I'm reverting the changes in the decoding part so that we can read our own datamatrices